task-management-agents/pyproject.toml
[project]
name = "task-management-agents"
version = "0.1.0"
description = "Multi-agent AI system for task, calendar, notes, and email management"
requires-python = ">=3.11"
dependencies = [
    # Google ADK & MCP
    "google-adk>=1.0.0",
    "mcp>=1.0.0",

    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",

    # Google Cloud services
    "google-cloud-firestore>=2.16.0",
    "google-api-python-client>=2.120.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
